--HG--
branch : sse2

+++ code/nel/include/nel/3d/driver.h
@@ -182,7 +182,7 @@ public:
 							IDriver();
 	virtual					~IDriver();
 
-	virtual bool			init(uint windowIcon = 0, emptyProc exitFunc = 0) = 0;
+	virtual bool			init(uintptr_t windowIcon = 0, emptyProc exitFunc = 0) = 0;
 
 	/// Deriver should calls IDriver::release() first, to destroy all driver components (textures, shaders, VBuffers).
 	virtual bool			release();
@@ -1341,7 +1341,7 @@ public:
 	 *	NB: if implementation does not support it, 0 may be returned. OpenGL ones return the Texture ID.
 	 *	NB: unlike isTextureExist(), this method is not thread safe.
 	 */
-	virtual	uint			getTextureHandle(const ITexture&tex) = 0;
+	virtual	uintptr_t		getTextureHandle(const ITexture&tex) = 0;
 
 	// see if the Multiply-Add Tex Env operator is supported (see CMaterial::Mad)
 	virtual	bool			supportMADOperator() const = 0;
